% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delim.R
\name{format_numeric}
\alias{format_numeric}
\title{Format numeric values for delimited output}
\usage{
format_numeric(x, digits = NULL, scientific = FALSE)
}
\arguments{
\item{x}{Numeric vector to format}

\item{digits}{Number of decimal places}

\item{scientific}{Whether to use scientific notation}
}
\value{
Character vector of formatted numbers
}
\description{
Format numeric values for delimited output
}
\keyword{internal}
