% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans-utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null Coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value to check}

\item{y}{Value to return if x is NULL}
}
\value{
x if not NULL, otherwise y
}
\description{
Internal utility function that returns the first non-NULL value.
}
\keyword{internal}
