% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans-utils.R
\name{knit_print.model_matrix}
\alias{knit_print.model_matrix}
\title{Format method for model_matrix objects in knitr}
\usage{
knit_print.model_matrix(x, ...)
}
\arguments{
\item{x}{A model_matrix object}

\item{...}{Additional arguments passed to knitr::kable}
}
\description{
Format method for model_matrix objects in knitr
}
