% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed.r
\name{knit_print.package_table}
\alias{knit_print.package_table}
\title{Custom knit_print method for package_table objects}
\usage{
knit_print.package_table(x, ...)
}
\arguments{
\item{x}{A package_table object}

\item{...}{Additional arguments passed to knitr::kable}
}
\description{
Custom knit_print method for package_table objects
}
