% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_values.r
\name{knit_print.scaled_values}
\alias{knit_print.scaled_values}
\title{Custom knit_print method for scaled_values objects}
\usage{
knit_print.scaled_values(x, ...)
}
\arguments{
\item{x}{A scaled_values object}

\item{...}{Additional arguments passed to knitr}
}
\description{
Custom knit_print method for scaled_values objects
}
