% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciiScan.R
\name{print.asciiScan_results}
\alias{print.asciiScan_results}
\title{Print method for asciiScan results}
\usage{
\method{print}{asciiScan_results}(x, ...)
}
\arguments{
\item{x}{Result object from asciiScan}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for asciiScan results
}
