% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateColnames.R
\name{print.colname_validation}
\alias{print.colname_validation}
\title{Print Method for Column Name Validation Results}
\usage{
\method{print}{colname_validation}(x, ...)
}
\arguments{
\item{x}{A colname_validation object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for Column Name Validation Results
}
