% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_df.R
\name{print.list_df}
\alias{print.list_df}
\title{Print method for list_to_df output}
\usage{
\method{print}{list_df}(x, ...)
}
\arguments{
\item{x}{Object of class list_df}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for list_to_df output
}
\keyword{internal}
