% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_data.R
\name{print.loaded_files}
\alias{print.loaded_files}
\title{Print method for loaded_files objects}
\usage{
\method{print}{loaded_files}(x)
}
\arguments{
\item{x}{Character, file directory}
}
\description{
Print method for loaded_files objects
}
