% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans-utils.R
\name{print.model_matrix}
\alias{print.model_matrix}
\title{Print method for model_matrix objects}
\usage{
\method{print}{model_matrix}(x, ...)
}
\arguments{
\item{x}{A model_matrix object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for model_matrix objects
}
