% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed.r
\name{print.package_table}
\alias{print.package_table}
\title{Print method for package_table objects}
\usage{
\method{print}{package_table}(x, ...)
}
\arguments{
\item{x}{A package_table object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for package_table objects
}
