% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_values.r
\name{print.scaled_values}
\alias{print.scaled_values}
\title{Print method for scaled_values objects}
\usage{
\method{print}{scaled_values}(x, ...)
}
\arguments{
\item{x}{A scaled_values object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for scaled_values objects
}
